requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pinecone-client==2.2.4
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.27
pypdf==4.0.1
google-generativeai==0.3.2
pydantic==2.6.1
typing-extensions==4.9.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
